refactor: rename qbd models to `qbd-*`
