+++ src/app/core/services/qbd/qbd-core/qbd-workspace.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
 import { QBDOnboardingState } from 'src/app/core/models/enum/enum.model';
-import { QBDWorkspace } from 'src/app/core/models/qbd/db/workspaces.model';
+import { QBDWorkspace } from 'src/app/core/models/qbd/db/qbd-workspace.model';
 import { Cacheable } from 'ts-cacheable';
 import { StorageService } from '../../common/storage.service';
 import { ApiService } from '../../common/api.service';
