+++ src/app/core/services/qbd/qbd-mapping/qbd-mapping.service.spec.ts
@@ -4,7 +4,7 @@ import { QbdMappingService } from './qbd-mapping.service';
 import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
 import { environment } from 'src/environments/environment';
 import { MappingState } from 'src/app/core/models/enum/enum.model';
-import { Mapping, MappingPost, MappingResponse, MappingStats } from 'src/app/core/models/qbd/db/mapping.model';
+import { Mapping, MappingPost, MappingResponse, MappingStats } from 'src/app/core/models/qbd/db/qbd-mapping.model';
 import { QBDExportSettingResponse, QBDExportSettingResponse2 } from 'src/app/integrations/qbd/qbd-shared/qbd-export-setting/qbd-export-setting.fixture';
 
 describe('QbdMappingService', () => {
