+++ src/app/integrations/qbd/qbd-main/qbd-dashboard/qbd-dashboard.fixture.ts
@@ -1,5 +1,5 @@
 import { QBDScheduleFrequency } from "src/app/core/models/enum/enum.model";
-import { QbdExportTriggerResponse, QbdAccountingExportDownload, QbdExportTriggerGet } from "src/app/core/models/qbd/db/iif-logs.model";
+import { QbdExportTriggerResponse, QbdAccountingExportDownload, QbdExportTriggerGet } from "src/app/core/models/qbd/db/qbd-iif-logs.model";
 import { QBDAdvancedSettingsGet } from "src/app/core/models/qbd/qbd-configuration/advanced-setting.model";
 
 export const getQbdAccountingExports: QbdExportTriggerResponse = {
